// src/Interfaces/IRequestConfig.ts
/**
 *
 */
export interface IRequestConfig {
    url: string;
    method?: Method;
    data?: any;
    headers?: any;
}

/**
 *
 */
export type Method =
  | 'get' | 'GET'
  | 'delete' | 'DELETE'
  | 'head' | 'HEAD'
  | 'options' | 'OPTIONS'
  | 'post' | 'POST'
  | 'put' | 'PUT'
  | 'patch' | 'PATCH'
  | 'purge' | 'PURGE'
  | 'link' | 'LINK'
  | 'unlink' | 'UNLINK'